Add social links to header
Duplicate social media links (formerly found in the footer only) so they
appear below the site logo in the header. This change is meant to
increase traffic to the Instagram page.

Additionally, remove some specific language from the home page copy that
referred to a situation from a few weeks ago, since it is no longer
relevant.