Remove incorrectly-scoped global style
Undo a recent style change to list items. A margin was added to list
items in the `main` content section, but this is too broadly applicable.

The actual markup that this change was added for has since been removed,
and therefore the style rule isn't needed at this time. In the future,
it will be easier to redo this change if it becomes necessary, with a
better implementation.